% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A017_addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\title{addAlpha}
\usage{
addAlpha(col, alpha = 1)
}
\arguments{
\item{col}{a colour}

\item{alpha}{a transparency from 0 to 1}
}
\value{
a colour
}
\description{
Takes a colour and adds transparency
}
\examples{
pruneStart(1:10,4)
}
