% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B003_applyShortcuts.R
\name{applyShortcuts}
\alias{applyShortcuts}
\title{applyShortcuts}
\usage{
applyShortcuts(inputList, useGrid = FALSE)
}
\arguments{
\item{inputList}{a parameter list with names}

\item{useGrid}{determines whether grid is used or not}
}
\description{
Takes a list with names that use shortcuts and replaces them with names recogised by R's graphics systems.
}
\examples{
print(1+1)
}
