% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A019_areaCalculations.R
\name{areaTriangle}
\alias{areaTriangle}
\title{areas}
\usage{
areaTriangle(x1, x2, x3, y1, y2, y3)
}
\arguments{
\item{x1, x2, x3, y1, y2, y3}{co-ordinates}
}
\value{
Returns area of triangle
}
\description{
functions to calculate the size of a triangle
}
\examples{
areaTriangle(0.5,0,1,1,0,0)
}
