% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D007_artdata.R
\docType{class}
\name{artdata}
\alias{artdata}
\alias{initialize,artdata-method}
\alias{+,artwork,artdata-method}
\alias{+,layer,artdata-method}
\alias{+,section,artdata-method}
\alias{+,component,artdata-method}
\alias{-,component,artdata-method}
\alias{-,section,artdata-method}
\alias{-,layer,artdata-method}
\alias{-,artwork,artdata-method}
\title{artdata}
\usage{
\S4method{initialize}{artdata}(.Object, pprefix = NULL, ..., artdata = list())

artdata(..., artdata = list())

\S4method{+}{artwork,artdata}(e1, e2)

\S4method{+}{layer,artdata}(e1, e2)

\S4method{+}{section,artdata}(e1, e2)

\S4method{+}{component,artdata}(e1, e2)

\S4method{-}{component,artdata}(e1, e2)

\S4method{-}{section,artdata}(e1, e2)

\S4method{-}{layer,artdata}(e1, e2)

\S4method{-}{artwork,artdata}(e1, e2)
}
\arguments{
\item{...}{extra parameters to be added to the data slot}

\item{artdata}{a list passed to the data slot}
}
\description{
An S4 class for storing parameters intended for art objects, object functions and plotting functions.
}
\details{
This class is a convenient way of adding parameters to components, layers, sections, scenes NOTYET and artwork objects.
}
\section{Slots}{

\describe{
\item{\code{artdata}}{a list of parameters}
}}

\examples{
print(1+1)
}
