% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D007_artp.R
\docType{class}
\name{artp}
\alias{artp}
\alias{initialize,artp-method}
\alias{+,artwork,artp-method}
\alias{+,layer,artp-method}
\alias{+,section,artp-method}
\alias{+,component,artp-method}
\alias{-,component,artp-method}
\alias{-,section,artp-method}
\alias{-,layer,artp-method}
\alias{-,artwork,artp-method}
\title{artp}
\usage{
\S4method{initialize}{artp}(.Object, pprefix = NULL, ..., artp = list())

artp(..., artp = list())

\S4method{+}{artwork,artp}(e1, e2)

\S4method{+}{layer,artp}(e1, e2)

\S4method{+}{section,artp}(e1, e2)

\S4method{+}{component,artp}(e1, e2)

\S4method{-}{component,artp}(e1, e2)

\S4method{-}{section,artp}(e1, e2)

\S4method{-}{layer,artp}(e1, e2)

\S4method{-}{artwork,artp}(e1, e2)
}
\arguments{
\item{...}{extra parameters to be added to the p slot}

\item{artp}{a list passed to the p slot}
}
\description{
An S4 class for storing parameters intended for art objects, object functions and plotting functions.
}
\details{
This class is a convenient way of adding parameters to components, layers, sections, scenes NOTYET and artwork objects.
}
\section{Slots}{

\describe{
\item{\code{artp}}{a list of parameters}
}}

\examples{
print(1+1)
}
