% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D006_artsettings.R
\docType{class}
\name{artsettings}
\alias{artsettings}
\alias{initialize,artsettings-method}
\alias{settings}
\alias{+,artwork,artsettings-method}
\title{artsettings}
\usage{
\S4method{initialize}{artsettings}(.Object, ..., settings = list())

settings(..., settings = list())

\S4method{+}{artwork,artsettings}(e1, e2)
}
\arguments{
\item{...}{extra parameters to be added to the settings slot}

\item{settings}{a list passed to the settings slot}
}
\description{
An S4 class for storing setting intended for artwork objects.
}
\details{
This class is a convenient way of adding settings to artwork objects.
}
\section{Slots}{

\describe{
\item{\code{settings}}{a list of settings}
}}

\examples{
print(1+1)
}
