% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D008_artstyle.R
\docType{class}
\name{artstyle}
\alias{artstyle}
\alias{initialize,artstyle-method}
\alias{+,artwork,artstyle-method}
\alias{+,section,artstyle-method}
\alias{+,layer,artstyle-method}
\alias{+,component,artstyle-method}
\alias{-,artwork,artstyle-method}
\alias{-,section,artstyle-method}
\alias{-,layer,artstyle-method}
\alias{-,component,artstyle-method}
\title{artstyle}
\usage{
\S4method{initialize}{artstyle}(.Object, ..., artstyle = list())

artstyle(...)

\S4method{+}{artwork,artstyle}(e1, e2)

\S4method{+}{section,artstyle}(e1, e2)

\S4method{+}{layer,artstyle}(e1, e2)

\S4method{+}{component,artstyle}(e1, e2)

\S4method{-}{artwork,artstyle}(e1, e2)

\S4method{-}{section,artstyle}(e1, e2)

\S4method{-}{layer,artstyle}(e1, e2)

\S4method{-}{component,artstyle}(e1, e2)
}
\arguments{
\item{...}{extra parameters to be added to the artstyle slot}

\item{artstyle}{a list passed to the artstyle slot}
}
\description{
An S4 class for storing parameters intended for par and gpar.
}
\details{
This class is a convenient way of adding styles to components, layers, sections, scenes NOTYET and artwork objects.
}
\section{Slots}{

\describe{
\item{\code{artstyle}}{a list of parameters}
}}

\examples{
print(1+1)
}
