% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C009_artwork.R
\docType{class}
\name{artwork}
\alias{artwork}
\alias{initialize,artwork-method}
\alias{+,artwork,list-method}
\alias{-,artwork,list-method}
\title{artwork}
\usage{
\S4method{initialize}{artwork}(
  .Object,
  ...,
  name = "artwork",
  pages = list(),
  deviceSettings = list(),
  format = "screen",
  folder = "",
  frameFolder = "",
  width = 4,
  height = 4,
  mode = 0,
  useGrid = FALSE,
  useShowText = FALSE,
  viewports = list(),
  style = list(mai = c(0, 0, 0, 0), xaxt = "n", yaxt = "n", yaxs = "i", xaxs = "i", mgp
    = c(0, 0, 0), oma = c(0, 0, 0, 0), plt = c(0, 1, 0, 1)),
  data = list(artBorder_col = "white", artBorder_border = "black"),
  units = list(),
  fonts = "",
  classes = c(artBorder = "rect"),
  psdata = list(),
  framerate = 10
)

\S4method{+}{artwork,list}(e1, e2)

\S4method{-}{artwork,list}(e1, e2)

artwork(...)
}
\arguments{
\item{name, format, deviceSettings, width, height, mode, folder, frameFolder, useGrid, useShowText, viewports, fonts, style, p, units, classes, psdata, framerate}{parameters for the constructor new("artwork") go directly into the relevant slots}

\item{pages, ...}{These parameters are merged into the pages slot.}
}
\description{
An S4 class to represent an entire graphic, artwork or animation;
}
\details{
The default artwork is a single frame image, displayed on screen with no margins, and no content.

The possible values of mode are:
\itemize{
\item mode=-2 All functions like component functions, layer functions etc are run and the resulting artwork is returned.
\item mode=-1 As mode -2, but only the maximal co-ordinates for xlim and ylim are returned.
\item mode=0 All frames are plotting on a single device, on page 1 of the file where appropriate
\item mode=1 On devices with multiple pages, seperate frames are on seperate pages.
\item mode=2 Seperate frames are saved in seperate files
\item mode=3 As mode 2, and ffmpeg is used to create a video afterwards.
}
}
\section{Slots}{

\describe{
\item{\code{pages}}{A list of pages contained in the layer}

\item{\code{name}}{A name for the file if saved to a single file}

\item{\code{format}}{Can be "screen", "pdf", "jpg", "bitmap", "cairoPDF", "png" or "cairoPNG".}

\item{\code{deviceSettings}}{A list of other device settings not covered by existing slots.}

\item{\code{width}}{width of the artwork}

\item{\code{height}}{height of the artwork}

\item{\code{mode}}{controls the meaning of frames and how animations are handled. See details.}

\item{\code{folder}}{location for the artwork to be saved if saved to a file}

\item{\code{frameFolder}}{location for frames to be saved if doing animations with mode 3}

\item{\code{useGrid}}{logical. if TRUE grid graphics are used instead of base graphics}

\item{\code{useShowText}}{logical to use showText in devices that support it.}

\item{\code{viewports}}{WHAT IS THIS?? DELETE????}

\item{\code{fonts}}{a character vector of font names.}

\item{\code{style}}{A list of parameters inherited by all pages, sections, layers and components in the artwork}

\item{\code{data}}{A list of parameters inherited by all pages, sections, layers and components in the artwork}

\item{\code{units}}{A list of units inherited by all pages, sections, layers and components in the artwork}

\item{\code{classes}}{A list that connects component types to plotting functions.}

\item{\code{psdata}}{NOT USED ANYWHERE WHAT IS THIS DELETE IT?????????}

\item{\code{framerate}}{A number used for animations (mode 3)}
}}

\examples{
print(1+1)
}
