% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A020_bezierInterpolate4Points.R
\name{bezierInterpolate4Points}
\alias{bezierInterpolate4Points}
\title{bezierInterpolate4Points}
\usage{
bezierInterpolate4Points(x, y, angles, tseq, tanInf = 100)
}
\arguments{
\item{x, y}{co-ordinates of the start and end points.}

\item{angles}{the first angle is the from the start and the second is from the end.}

\item{tseq}{sequence of parameters for the bezier curve}

\item{tanInf}{value at which tan of angle is considered large enough to be infinite. I think this avoids dividing by zero? or smoother vertical lines???}
}
\value{
Returns a curve of points. See bezier() in the bezier package for more details.
}
\description{
Performs a bezier interpolation based on four points. Instead of requiring the middle two points, the angle of vectors from the start to point 2 and from the end to point 3 are required.
}
\examples{
print(1+1)
}
