% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D005_c2ccf.R
\docType{class}
\name{c2ccf}
\alias{c2ccf}
\alias{initialize,c2ccf-method}
\alias{c2c_cf}
\alias{+,component,c2ccf-method}
\alias{-,component,c2ccf-method}
\title{c2ccf}
\usage{
\S4method{initialize}{c2ccf}(.Object, ..., functionList = as.character())

c2c_cf(functions = c(), ...)

\S4method{+}{component,c2ccf}(e1, e2)

\S4method{-}{component,c2ccf}(e1, e2)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing component functions that accept components.
}
\details{
This class is a convenient way of adding functions to component objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
