% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D002_calculateXYScales.R
\name{calculateXYscale}
\alias{calculateXYscale}
\title{calculateXYscale}
\usage{
calculateXYscale(
  art,
  preserveAspectRatio = FALSE,
  aspectRatio = 1,
  innerMargin = 0.04
)
}
\arguments{
\item{art}{an artwork object}

\item{preserveAspectRatio}{logical}

\item{aspectRatio}{numerical}

\item{innerMargin}{numerical}
}
\value{
returns a list containing two vectors x and y which contain minimum and maximum values.
}
\description{
A function to calculate the minimum/maximum co-ordinates of an artwork  object
}
\details{
This function is used by easyPlot to automatically calculate xlim and ylim if used on components or layers, or a list of such objects.
It runs all the functions in the artwork, so it will double the length of time it takes to plot an artwork. See easyPlot for more details.
}
\examples{
print(1+1)
}
