% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B006_calibrateParameters.R
\name{calibrateParameters}
\alias{calibrateParameters}
\title{calibrateParameters}
\usage{
calibrateParameters(
  styleList,
  simplify = TRUE,
  fun = NULL,
  style = TRUE,
  classes = as.character(),
  vp = list(x = 0, y = 0),
  units = list(),
  xscale = 0:1,
  yscale = 0:1,
  useGrid = FALSE,
  additions = list()
)
}
\arguments{
\item{styleList}{a list of parameters}

\item{simplify}{logical maybe always TRUE}

\item{fun}{the function}

\item{style}{logical determines if the parameters are for par/gpar (TRUE) or a plotting function (FALSE)}

\item{classes}{classes for plot functions}

\item{vp}{the co-ordinates for the parent section}

\item{units}{determines the units for co-ordinates}

\item{xscale, yscale}{vector of min and max for native units}

\item{useGrid}{logical}
}
\value{
Returns a named list
}
\description{
Takes a list of parameters and tunes them to the requirements. This includes removing shortcuts and calculations for combining units.
}
\examples{
print(1+1)
}
