% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A016_polarConversions2.R
\name{cart2polar}
\alias{cart2polar}
\title{cart2polar}
\usage{
cart2polar(
  x = 0,
  y = 0,
  x0 = NULL,
  y0 = NULL,
  x1 = NULL,
  y1 = NULL,
  lineSegments = FALSE,
  forcePositive = FALSE
)
}
\arguments{
\item{x, y, x0, y0, x1, y1}{numerical. See details.}

\item{lineSegments}{logical. See details.}

\item{forcePositive}{defaults to FALSE. Setting this to true ensures the bearings are between 0 and 2*pi instead of from -pi to pi.}
}
\value{
Returns a list with r and bearing co-ordinates.
}
\description{
Converts polar co-ordinates into cartesian co-ordinates.
}
\details{
If none of the below apply (the default), the input vector is (x,y). i.e. from the origin to x,y.

x and y are ignored if all of x0,y0,x1,y1 are provided and the input vectors are (x1-x0,y1-y0). i.e. from (x0,y0) to (x1,y1). x1 and y1 are ignored otherwise.

if all of x0,y0,x1,y1 are NULL, only x and y are used. If lineSegments is FALSE, the default behaviour applies. If TRUE, the vectors are along the line given by x and y. i.e. from (x\link{1},y\link{1}) to (x\link{2},y\link{2}), from (x\link{2},y\link{2}) to (x\link{3},y\link{3}), etc.
}
\examples{
print(1+1)
}
\seealso{
Other polarConversion: 
\code{\link{makePolar}()},
\code{\link{polar2cart}()}
}
\concept{polarConversion}
