% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G003_celticKnots.R
\name{celticKnots}
\alias{celticKnots}
\alias{startCelticRibbon}
\alias{celticBezierCurve}
\alias{loopToRibbon}
\alias{reorderCelticKnotComponents}
\alias{makeCelticCurve}
\title{Celtic Knots}
\usage{
startCelticRibbon(layer)

celticBezierCurve(layer)

loopToRibbon(layer)

reorderCelticKnotComponents(layer)

makeCelticCurve(
  lineOrder = c("A", "D", "H", "L", "M", "I", "F", "B", "C", "G", "K", "N", "J", "E"),
  angles = data.frame(angleNumber = c(1:5, 7:12, 14), inAngle1 = pi * c(1/2, 1/4, 0,
    -1/4, 1, 3/4, -1/2, -1/4, 0, 1/4, 1, -3/4), inAngle2 = pi * c(1/4, 0, -1/4, -1/2,
    3/4, 1, -1/4, 0, 1/4, 1/2, -3/4, 1), stringsAsFactors = FALSE),
  basis = data.frame(names = LETTERS[1:14], x = c(1, 1, 3, 3, 5, 5, 6, 6, 7, 7, 9, 9,
    11, 11), y = c(1, 7, 3, 5, 3, 5, 1, 7, 3, 5, 3, 5, 1, 7), stringsAsFactors = FALSE),
  knotOrder = c(1, 3, 5, 7, 9, 11, 13, 2, 4, 6, 8, 10, 12, 14),
  knotFillColour = list("white"),
  knotBorderColour = "black",
  clockWiseCorners = c(1:5, 8:12),
  flipCorners = c(13, 14),
  distance = 0.2,
  knotMode = "DEFAULT",
  ...
)
}
\arguments{
\item{...}{parameters passed to the layer and the layer function object. See "details" for more information.}
}
\value{
a celtic knot
}
\description{
Functions to draw Celtic knots.
}
\examples{
print(1+1)
}
\seealso{

}
