% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G001_chaosGame.R
\name{ChaosGame}
\alias{ChaosGame}
\alias{makeChaosGame}
\alias{runChaosGame}
\title{Chaos Game}
\usage{
makeChaosGame(type = "points", ...)

runChaosGame(
  inputx = 0,
  inputy = 0,
  x = NULL,
  y = NULL,
  attractor = list(x = c(-1, -1, 1, 1), y = c(-1, 1, 1, -1)),
  decisions = NULL,
  runTime = 1000,
  sampleSize = NULL,
  reuseInput = FALSE,
  decisionMode = 1
)
}
\arguments{
\item{type}{The name(s) of the plotting function(s) used to plot the component. Defaults to "points". This can also include the names of classes linked to a plotting function.}

\item{...}{parameters passed to the component and the component function runChaosGame. See "details" for more information.}
}
\description{
runChaosGame calculates the points. makeChaosGame makes a component that calculates and plots the points.
}
\examples{
art=makeChaosGame(decisionMode=2,cex=0.1,pch=21,runTime=10000)
#easyPlot(art,preserveAspectRatio=TRUE)
}
