% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J007_checkPatterns.R
\name{checkPatterns}
\alias{checkPatterns}
\alias{checkPattern1}
\alias{checkPattern2}
\title{Check Patterns}
\usage{
checkPattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

checkPattern2(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)
}
\arguments{
\item{x0, x1, y0, y1}{co-ordinates for the top, bottom, left and right of the square.}

\item{backgroundBorder, patternBorder}{colour for the lines in the background and foreground.}

\item{foreground, background}{colours for the foreground and background.}

\item{...}{parameters passed to the component.}
}
\value{
return a single component
}
\description{
Functions to create chequered patterns
}
\details{
Two functions that make checkered patterns.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{makePatternBlock}()},
\code{\link{patternBasics}},
\code{\link{patternNames}()},
\code{\link{squarePatterns}},
\code{\link{stripePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
