% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E008_colourComponents.R
\name{colourComponents}
\alias{colourComponents}
\title{colourComponents}
\usage{
colourComponents(layer)
}
\arguments{
\item{layer}{input layer}
}
\value{
Returns the original layer with modified colours for the components
}
\description{
Takes a layer and uses the componentColours entry in the data slot to colour the components.
}
\examples{
print(1+1)
}
