% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourScheme.R
\name{colourScheme}
\alias{colourScheme}
\title{colourScheme}
\usage{
colourScheme(
  numberOfColours = 10,
  style = "BW",
  colours = NULL,
  bias = 1,
  space = c("rgb", "Lab"),
  alpha = FALSE,
  interpolate = c("linear", "spline"),
  reverse = FALSE,
  discrete = TRUE
)
}
\arguments{
\item{numberOfColours}{The number of colours returned. Defaults to 10.}

\item{style}{The palette or function to use. Used when colours is NULL. Defaults to "BW" for black and white.}

\item{colours}{A starting set of colours that will be interpolated.}

\item{bias}{used by colorRampPalette}

\item{space}{used by colorRampPalette}

\item{alpha}{used by colorRampPalette}

\item{interpolate}{used by colorRampPalette}

\item{reverse}{Reverses the final colour order. Defaults to FALSE}

\item{discrete}{Defaults to TRUE}
}
\value{
Returns a vector of colours of the given length.
}
\description{
Creates a single function that can create colour schemes using many different packages with the option of interpolation.
}
\examples{
colourScheme()
colourScheme(25,"rainbow")
}
