% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C002_componentFunctions.R
\docType{class}
\name{componentfunctions}
\alias{componentfunctions}
\alias{initialize,componentfunction-method}
\alias{componentfunction}
\alias{runComponentFunction}
\title{componentfunctions}
\usage{
\S4method{initialize}{componentfunction}(
  .Object,
  fun = function(x) {     x },
  repeats = 1,
  returnComponents = TRUE,
  frames = integerSet()
)

componentfunction(...)

runComponentFunction(input, inputFun, frameNumber)
}
\description{
An S4 class for functions that operate on components or their parameters
}
\details{
action.data, action.comp, build.data, build.comp, display.data and display.comp are convenient ways to add componentfunctions to components.
}
\section{Slots}{

\describe{
\item{\code{fun}}{the name of a function or an function.}

\item{\code{repeats}}{numeric. The number of times the function runs per frame when it runs.}

\item{\code{returnComponents}}{logical. If TRUE the function in fun accepts and returns component objects. If FALSE it accepts parameters like a normal R function and returns a named list of parameters.}

\item{\code{frames}}{an integerSet object. Controls on which frames the function runs. The default behaviour is that it runs on all frames.}
}}

\examples{
print(1+1)
}
