% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E002_defaultArtSettings.R
\name{defaultArtSettings}
\alias{defaultArtSettings}
\title{defaultArtSettings}
\usage{
defaultArtSettings()
}
\value{
Returns a list including artClasses, artp, and artstyle.
}
\description{
Returns a list of default settings and parameters for easyPlot.
}
\examples{
print(1+1)
}
