% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E003_easyPlot.R
\name{easyPlot}
\alias{easyPlot}
\title{easyPlot}
\usage{
easyPlot(
  art,
  border = 0.1,
  useGrid = FALSE,
  format = "screen",
  width = NULL,
  height = NULL,
  name = NULL,
  folder = NULL,
  frameFolder = NULL,
  xlim = NULL,
  ylim = NULL,
  widthUnit = "in",
  heightUnit = "in",
  deviceSettings = list(res = 150, antialias = "cleartype"),
  bg = NA,
  borderLine = "black",
  preserveAspectRatio = FALSE,
  .aspectRatio = 1,
  plotArt = TRUE,
  mode = NULL,
  frames = 1,
  framesToPlot = NULL,
  framerate = 10,
  innerMargin = 0.04,
  ...
)
}
\arguments{
\item{art}{either a component, a list of components, a layer, a list of layers, a section, a page or an artwork.}

\item{border}{size of border(s)}

\item{useGrid}{logical}

\item{format}{string}

\item{width, height}{numerical}

\item{name}{string}

\item{folder}{string}

\item{frameFolder}{string}

\item{xlim, ylim}{numerical}

\item{widthUnit, heightUnit}{list of units}

\item{deviceSettings}{list of settings}

\item{bg}{background colour}

\item{borderLine}{border colour}

\item{preserveAspectRatio}{logical. see addBorder}

\item{.aspectRatio}{numerical #why .? see addBorder}

\item{plotArt}{logical. If FALSE simply returns the artwork.}

\item{mode}{numeric.}

\item{frames}{numeric.}

\item{framesToPlot}{numeric.}

\item{framerate}{numeric.}

\item{innerMargin}{numeric. see addBorder}

\item{...}{parameters passed to the final artwork using art=art+artdata(...)}
}
\value{
depends on plotArt and mode and art@mode.
}
\description{
Plots any art object and accepts parameters for a quick way to plot and modify art objects.
}
\examples{
print(1+1)
}
