% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D005_functionList.R
\docType{class}
\name{functionList}
\alias{functionList}
\alias{initialize,functionList-method}
\alias{+,component,functionList-method}
\alias{-,component,functionList-method}
\alias{+,layer,functionList-method}
\alias{-,layer,functionList-method}
\alias{+,section,functionList-method}
\alias{-,section,functionList-method}
\title{functionList}
\usage{
\S4method{initialize}{functionList}(
  .Object,
  ...,
  functionList = as.character(),
  frames = integerSet(),
  abcd = "a",
  repeats = 1,
  useData = TRUE,
  type = ""
)

functionList(functions = c(), ...)

\S4method{+}{component,functionList}(e1, e2)

\S4method{-}{component,functionList}(e1, e2)

\S4method{+}{layer,functionList}(e1, e2)

\S4method{-}{layer,functionList}(e1, e2)

\S4method{+}{section,functionList}(e1, e2)

\S4method{-}{section,functionList}(e1, e2)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing functions for components, layers, sections and so on.
}
\details{
functionList is the recommended way to add functions to art objects inside a make function.
Functions like action.data are another way to do this.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
