% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B002_grepRename.R
\name{grepRename}
\alias{grepRename}
\title{grepRename}
\usage{
grepRename(inputList, inputPattern, replaceString)
}
\arguments{
\item{inputList}{a named list}

\item{inputPattern}{the items to be renamed}

\item{replaceString}{the replacement}
}
\value{
Returns the original list with any matches renamed.
}
\description{
A function to rename the names in a named list using grep patterns.
}
\examples{
testInputList=list(col="red",.colour="blue",points_.colour="green",lines_col="yellow",lines_.colour="pink")
testOutputList=grepRename(testInputList,"\\\\.colour$","col")
expectedOutputList=list(col="blue",points_col="green",lines_col="pink")
}
