% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B004_inheritParameters.R
\name{inheritParameters}
\alias{inheritParameters}
\title{inheritParameters}
\usage{
inheritParameters(
  parentList,
  childList,
  classes = as.character(),
  applyShortcuts = TRUE,
  useGrid = FALSE
)
}
\arguments{
\item{parentList, childList}{lists of parameters. In case of a clash, the childList takes precedence.}

\item{classes}{list of classes}

\item{applyShortcuts}{CURRENTLY ALWAYS TRUE???}

\item{useGrid}{logical. Needed to apply shortcuts.}
}
\value{
Returns a list that combines parentList and childList, but resolves any conflicts in favour of the child.
}
\description{
Combines parameters from two objects, for example a component and its parent layer.
}
\examples{
print(1+1)
}
