% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A009_instagramColours.R
\name{instagramColours}
\alias{instagramColours}
\title{instagramColours}
\usage{
instagramColours()
}
\value{
Returns a vector of colour names
}
\description{
returns a colour palette
}
\examples{
instagramColours()
}
