% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H001_juliaPlots.R
\name{juliaPlots}
\alias{juliaPlots}
\alias{runJuliaPlot}
\alias{makeJuliaPlot}
\title{Julia Plots}
\usage{
runJuliaPlot(
  size = c(1000, 1000),
  xScale = c(-2, 2),
  yScale = c(-2, 2),
  juliaFunction = NULL,
  coefficient = complex(real = 0.276, imaginary = 0),
  maxIter = 100,
  maxBound = 4
)

makeJuliaPlot(...)
}
\arguments{
\item{size}{vector length 2 giving the dimensions of the matrix}

\item{xScale, yScale}{}

\item{coefficient}{The c parameter in z^2+c. Vector length 2 representing the real and imaginary parts of the number c.}

\item{maxIter}{The number of steps at which a point is considered to be bounded.}

\item{maxBound}{The distance at which a point is considered to have escaped to infinity.}

\item{...}{parameters passed to the component and the component function runJuliaPlot. See "details" for more information.}
}
\value{

}
\description{
Functions to run the calculations for Julia Plots.
}
\examples{
print(1+1)
}
