% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D010_l2llf.R
\docType{class}
\name{l2llf}
\alias{l2llf}
\alias{initialize,l2llf-method}
\alias{l2l_lf}
\alias{+,layer,l2llf-method}
\alias{-,layer,l2llf-method}
\title{l2llf}
\usage{
\S4method{initialize}{l2llf}(.Object, ..., functionList = as.character())

l2l_lf(functions = c(), ...)

\S4method{+}{layer,l2llf}(e1, e2)

\S4method{-}{layer,l2llf}(e1, e2)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing layer functions that accept layers.
}
\details{
This class is a convenient way of adding functions to layer objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
