% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C004_layerFunctions.R
\docType{class}
\name{layerfunctions}
\alias{layerfunctions}
\alias{initialize,layerfunction-method}
\alias{layerfunction}
\alias{runLayerFunction}
\alias{runLayerFunctions}
\title{layerfunctions}
\usage{
\S4method{initialize}{layerfunction}(
  .Object,
  fun = function(x) {     x },
  returnLayer = TRUE,
  frames = integerSet()
)

layerfunction(...)

runLayerFunction(input, inputFun, classes, frameNumber)

runLayerFunctions(
  input,
  classes = list(),
  getData = FALSE,
  frameNumber,
  abcd = "a"
)
}
\description{
An S4 class for functions that operate on layers or their parameters
}
\details{
action.layer, action.data, etc... are convenient ways to add layerfunctions to layers.
}
\section{Slots}{

\describe{
\item{\code{fun}}{the name of a function or an function.}

\item{\code{returnLayer}}{logical. If TRUE the function in fun accepts and returns layer objects. If FALSE it accepts parameters like a normal R function and returns a named list of parameters.}

\item{\code{frames}}{an integerSet object. Controls on which frames the function runs. The default behaviour is that it runs on all frames.}
}}

\examples{
print(1+1)
}
