% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C003_layer.R
\name{layers}
\alias{layers}
\alias{layer}
\alias{initialize,layer-method}
\alias{+,layer,component-method}
\alias{-,layer,component-method}
\title{layers}
\usage{
layer(...)

\S4method{initialize}{layer}(
  .Object,
  ...,
  style = list(),
  visible = TRUE,
  units = list(),
  p = list(),
  functions = list(),
  components = list(),
  layers = list(),
  frames = integerSet()
)

\S4method{+}{layer,component}(e1, e2)

\S4method{-}{layer,component}(e1, e2)

layer(...)
}
\arguments{
\item{components, ...}{These parameters are merged into the components slot.}

\item{layers, style, units, functions, visible}{parameters for the constructor new("layer") go directly into the relevant slots}

\item{frames}{If this is an integerSet, it goes directly into the relevant slot. If it is a numeric vector it is converted into an integerSet for you.}
}
\description{
An S4 class to represent a collection of single elements in a graphic;
}
\details{
The default layer made by the constructor completely empty, with no parameters, no components, no sublayers, no functions, no units. It is active and plotted on every frame.
}
\section{Slots}{

\describe{
\item{\code{components}}{A list of components contained in the layer}

\item{\code{layers}}{A list of layers contained in the layer}

\item{\code{units}}{A list of units inherited by components and layers in this layer}

\item{\code{p}}{a list of parameters inherited by components and layers in this layer}

\item{\code{style}}{a list of parameters inherited by components and layers in this layer}

\item{\code{visible}}{logical. If FALSE the layer isn't plotted.}

\item{\code{functions}}{list. Functions that act on the layer or the elements of p or the list of components.}

\item{\code{frames}}{an integerSet. The set of frames the layer can be displayed in, (assuming visible is TRUE).}
}}

\examples{
print(1+1)
}
