% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A003_loopIndex.R
\name{loopIndex}
\alias{loopIndex}
\title{loopIndex}
\usage{
loopIndex(i, n)
}
\arguments{
\item{i}{a vector}

\item{n}{a number}
}
\value{
Returns the vector i modulo n in the range 1 to n
}
\description{
Calculates i modulo n, but returns numbers in the range from 1 to n instead of 0 to n-1.
}
\examples{
loopIndex(1:10,4)
}
