% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E006_gallery.R
\name{makeGallery}
\alias{makeGallery}
\title{makeGallery}
\usage{
makeGallery(
  inputList,
  nRow,
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  x = unit(0, "npc"),
  y = unit(0, "npc"),
  xscale = NULL,
  yscale = NULL,
  useGlobalScale = FALSE,
  units = list(),
  preserveAspectRatio = FALSE
)
}
\arguments{
\item{nRow}{number of rows in the gallery.}

\item{width, height, x, y}{parameters for the final section}

\item{xscale, yscale}{a scale to use for each child section. If not provided these are calculated. See useGlobalScale.}

\item{useGlobalScale}{if FALSE child scaling calculations are independent. If TRUE, scales are adjusted so each child has the same scales, and all components are visible. Ignored if xscale and yscale are provided.}

\item{units}{provided to each child section}

\item{preserveAspectRatio}{logical. Used if scales are automatically calculated.}
}
\value{
Returns a section object.
}
\description{
Takes a list of components and/or layers and creates a section that contains a grid of sections which each contain one item from the original list.
}
\examples{
pruneStart(1:10,4)
}
