% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F002_image.R
\name{makeImage}
\alias{makeImage}
\alias{makeImageFromMatrix}
\title{makeImage}
\usage{
makeImage(
  type = "rasterImage",
  xleft = 0,
  xright = 1,
  ytop = 1,
  ybottom = 0,
  ...
)

makeImageFromMatrix(.useData = c(FALSE, TRUE), ...)
}
\arguments{
\item{type}{plotting function(s) used}

\item{xleft, xright, ytop, ybottom}{positioning. see rasterImage}

\item{...}{parameters passed to the component or the component functions imageColourScheme and matrixToColours.}
}
\value{
a component
}
\description{
Functions to make a component for the rasterImage function in the base graphics system.
makeImage assumes the input "image"is an RGB array, whilst makeImageFromMatrix converts a matrix into colours for you using matrixToColours.
}
