% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K003_mandalas.R
\name{makeMandalaStringArt}
\alias{makeMandalaStringArt}
\title{makeMandalaStringArt}
\usage{
makeMandalaStringArt(
  x = 0,
  y = 0,
  radii = c(1, 2, 3, 4, 5),
  numPoints = c(16, 16, 16, 16, 16),
  distances = c(3, 1, 1, 4),
  loop = rep(TRUE, 4),
  ...
)
}
\arguments{
\item{x, y}{centre of the circles}

\item{radii}{the size of each circle. Length of this parameter controls the number of circles.}

\item{numPoints}{vector. The number of points on each circle}

\item{distances, loop}{string art parameters}
}
\value{
Returns a component that represents a piece of stringArt.
}
\description{
Creates string art whose points are on circles and the lines are between the circles.
}
\examples{
print(1+1)
}
\seealso{
Other stringArt: 
\code{\link{makeShapeStringArt}()},
\code{\link{makeStringArt}()},
\code{\link{stringArt}}
}
\concept{stringArt}
