% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M004_maurerRose.r
\name{makeMaurerRose}
\alias{makeMaurerRose}
\title{maurerRose}
\usage{
makeMaurerRose(
  petalNumber = 2,
  degreesParameter = 39,
  lent = 361,
  tmin = 0,
  tmax = 2 * pi,
  ...
)
}
\arguments{
\item{petalNumber, degreesParameter}{maurer rose parameters.}

\item{lent, tmin, tmax}{passed to makeTSpace()}

\item{...}{parameters passed to makePolar}
}
\value{
Returns a component
}
\description{
Creates a Maurer rose.
}
\examples{
pruneStart(1:10,4)
}
