% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F003_mesh.R
\name{makeMesh}
\alias{makeMesh}
\alias{runMesh}
\title{makeMesh}
\usage{
runMesh(x, y, z, faces)

makeMesh(type = "polygon", ...)
}
\arguments{
\item{x, y, z}{co-ordinates for vertices}

\item{faces}{list of faces. Each face is numerical vector of indices that specify a sequence of vertices from x,y and z.}
}
\description{
Functions to produces meshes from a list of vertices
runMesh performs calculations. makeMesh makes a component or list of components.
}
\concept{makeMesh
makeMesh()}
