% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M001_parametric.R
\name{makeParametric}
\alias{makeParametric}
\alias{runParametric}
\alias{makeTSpace}
\title{makeParametric}
\usage{
runParametric(
  xFun = function(x) {     sin(x * 2 * pi) },
  yFun = function(x) {     sin(x * 4 * pi) },
  parameterSpace
)

makeTSpace(tmin = 0, tmax = 1, lent = 101)

makeParametric(type = "lines", ...)
}
\arguments{
\item{xFun, yFun}{functions or names of functions.}

\item{parameterSpace}{the set of parameters used.}

\item{tmin, tmax, lent}{used when parameterSpace isn't given. seq(tmin,tmax,lent) is then used as a parameterSpace.}

\item{...}{parameters passed to the component and the component functions runParametric and runTSpace. See "details" for more information.}
}
\description{
Functions to calculate the sequence of points given by functions x(t),y(t) for a vector of parameters t.
runParametric calculates the points. makeParametric makes a component that calculates and plots the output. runTSpace creates a set of parameters if one isn't provided.
}
\concept{makeParametric
easyPlot(makeParametric())}
