% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J002_makePatternBlock.R
\name{makePatternBlock}
\alias{makePatternBlock}
\alias{runPatternBlock}
\title{makePatternBlock}
\usage{
runPatternBlock(input)

makePatternBlock(...)
}
\arguments{
\item{...}{parameters passed to the component and the function runPatternBlock. See details.}
}
\value{
makePatternBlock creates a component whilst runPatternBlock takes in a component and returns a list of components.
}
\description{
Functions to create a block of the same pattern repeated over and over.
}
\details{
makePatternBlock creates an entire block of patterns.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{checkPatterns}},
\code{\link{patternBasics}},
\code{\link{patternNames}()},
\code{\link{squarePatterns}},
\code{\link{stripePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
