% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M002_polar.R
\name{makePolar}
\alias{makePolar}
\alias{runPolar}
\title{Polar Co-ordinates}
\usage{
runPolar(
  x = 0.5,
  y = 0.5,
  distanceFunction = function(x) {     x/2 },
  bearingFunction = function(x) {     x * 2 * pi },
  parameterSpace = seq(0, 1, length.out = 100),
  rotation = 0,
  numLines = 1
)

makePolar(type = "lines", ...)
}
\arguments{
\item{x, y}{the center polar co-ordinates are drawn from.}

\item{distanceFunction}{a function that returns distances from x,y}

\item{bearingFunction}{a function that returns bearings from x,y}

\item{parameterSpace}{a vector of parameters passed to the distance and bearing functions.}

\item{rotation}{an additional rotation. Defaults to 0.}

\item{numLines}{divides the answer into a number of lines seperated by NAs for use with lines().}

\item{type}{the type/graphics function of the component.}

\item{...}{parameters passed to the component and the component functions. See "details" for more information.}
}
\value{
runPolar returns a list with x and y components. makePolar returns a component.
}
\description{
Functions to convert polar co-ordinates into cartesean co-ordinates and to make a component that uses functions that output polar co-ordinates.
}
\examples{
print(1+1)
}
\seealso{


Other polarConversion: 
\code{\link{cart2polar}()},
\code{\link{polar2cart}()}
}
\concept{polarConversion}
