% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K004_shapeStringArt.R
\name{makeShapeStringArt}
\alias{makeShapeStringArt}
\title{makeShapeStringArt}
\usage{
makeShapeStringArt(
  x = c(0, 1, 1),
  y = c(0, 0, 1),
  focus = NULL,
  scales = (10:6)/10,
  distances = c(0, 1, 0, 1),
  loop = rep(TRUE, 4),
  ...
)
}
\arguments{
\item{x, y}{co-ordinates of the original shape.}

\item{focus}{co-ordiantes of the centre to scale from. NULL, the default, uses c(mean(x),mean(y)) as the focus point.}

\item{scales}{scaling factors for the shapes.}

\item{distances, loop}{string art parameters}
}
\value{
Returns a component that represents a piece of stringArt.
}
\description{
Creates string art whose points are on scaled variations of a starting shape and the lines are between the shapes.
}
\examples{
print(1+1)
}
\seealso{
Other stringArt: 
\code{\link{makeMandalaStringArt}()},
\code{\link{makeStringArt}()},
\code{\link{stringArt}}
}
\concept{stringArt}
