% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M005_spirals.r
\name{makeSpiral}
\alias{makeSpiral}
\title{spiral}
\usage{
makeSpiral(
  numLines = 1,
  x = 0.5,
  y = 0.5,
  maxRadius = 0.49,
  numPoints = 100,
  numTurns = 3.5,
  clockwise = FALSE,
  ...
)
}
\arguments{
\item{numLines}{the number of lines spiralling out from the center}

\item{x, y}{co-ordinates for the centre of the spiral}

\item{maxRadius}{the size of the spiral}

\item{numPoints}{the number of points. \link{CHECK per line or per spiral??}}

\item{numTurns}{the number of times a line goes around the centre of the spiral.}

\item{clockwise}{logical. Sets the direction or rotation. Defaults to FALSE / anti-clockwise.}

\item{...}{parameters passed to makePolar}
}
\value{
Returns a component
}
\description{
Creates a spiral.
}
\examples{
pruneStart(1:10,4)
}
