% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M006_spirograph.R
\name{makeSpirograph}
\alias{makeSpirograph}
\title{spirograph}
\usage{
makeSpirograph(
  discSizes = c(1, 0.4),
  discRotations = c(3, 25),
  tmin = 0,
  tmax = 2 * pi,
  lent = 1000,
  ...
)
}
\arguments{
\item{discSizes, discRotations}{spirograph parameters.}

\item{lent, tmin, tmax}{passed to makeTSpace()}

\item{...}{parameters passed to makePolar}
}
\value{
Returns a component
}
\description{
Creates a spirograph.
}
\examples{
pruneStart(1:10,4)
}
