% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K002_stringArt.R
\name{makeStringArt}
\alias{makeStringArt}
\alias{runStringArt}
\title{makeStringArt}
\usage{
runStringArt(stringArtObject)

makeStringArt(
  type = "segments",
  names,
  x,
  y,
  groups,
  lineGroups,
  distances,
  addPoints = FALSE,
  loop,
  ...
)
}
\arguments{
\item{stringArtObject}{used by runStringArt}

\item{names, x, y, groups, lineGroups, distances, addPoints=FALSE, loop}{parameters used to create a stringArtObject}

\item{...}{parameters passed to the component. See "details" for more information.}
}
\value{
runStringArt returns a list with x0,x1,y0,y1 co-ordinates. makeStringArt returns a component object.
}
\description{
Functions to calculate the co-ordinates of a piece of string art from a string art object and to make a component to represent a string art object.
}
\details{
For drawing string art.
}
\examples{
print(1+1)
}
\seealso{


Other stringArt: 
\code{\link{makeMandalaStringArt}()},
\code{\link{makeShapeStringArt}()},
\code{\link{stringArt}}
}
\concept{stringArt}
