% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F004_targets.R
\name{makeTarget}
\alias{makeTarget}
\title{makeTarget}
\usage{
makeTarget(
  x,
  y,
  maxRadius,
  numberOfColours,
  .useData = c(TRUE, FALSE),
  addColoursName = "componentColours",
  ...
)
}
\arguments{
\item{x, y}{co-ordinates for the centre of the targetColours}

\item{maxRadius}{size of largest circle}

\item{numberOfColours}{number of circles}

\item{.useData}{controls addColours and colourComponents.}

\item{...}{passed to layer@data and function objects}
}
\value{
Returns a layer with one component per circle.
}
\description{
Creates a layer containing a series of concentric circles.
}
\examples{
easyPlot(makeTarget(x=0,y=0,1,3)
}
