% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A013_matrixManipulation.R
\name{matrixManipulation}
\alias{matrixManipulation}
\alias{shiftMatrix}
\alias{neighbourhood}
\title{matrixManipulation}
\usage{
shiftMatrix(input, xshift = 0, yshift = 0)

neighbourhood(input, type = "VN", range = 1)
}
\arguments{
\item{input}{a matrix.}

\item{xshift, yshift}{amount to shift the matrix}

\item{type}{"N","VN" or "M".}

\item{range}{numeric. The size of the neighbourhood}
}
\value{
shiftMatrix returns a matrix and neighbourhood returns a array.
}
\description{
Functions to find the neighbouring values of matrix cells. Used by Cyclic Cellular Automata.
}
