% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D001_measuringObjects.R
\name{measuringObjects}
\alias{measuringObjects}
\alias{xScaleComponent}
\alias{yScaleComponent}
\alias{xScaleLayer}
\alias{yScaleLayer}
\alias{xScaleSection}
\alias{yScaleSection}
\title{measuringObjects}
\usage{
xScaleComponent(input, cmin = NULL, cmax = NULL)

yScaleComponent(input, cmin = NULL, cmax = NULL)

xScaleLayer(input, cmin = NULL, cmax = NULL)

yScaleLayer(input, cmin = NULL, cmax = NULL)

xScaleSection(input, cmin = NULL, cmax = NULL)

yScaleSection(input, cmin = NULL, cmax = NULL)
}
\arguments{
\item{input}{a component, layer or section depending on the function}

\item{cmin, }{cmax numeric. The current minimum or maximum to compare against. If the co-ordinates in the input do not exceed cmin or cmax, cmin and cmax will be returned instead, unless they are NULL.}
}
\value{
a vector of length two with the minimum and maximum co-ordinates.
}
\description{
Functions to find the minimum and maximum x or co-ordinates of components, layers or sections.
}
\examples{
print(1+1)
}
