% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A010_newFileName.R
\name{newFileName}
\alias{newFileName}
\title{newFileName}
\usage{
newFileName(folder, name, format)
}
\arguments{
\item{folder}{the address of the folder}

\item{name}{the non-numerical part of the file name}

\item{format}{the type of file used. (png, txt, etc.)}
}
\value{
Returns a new file name
}
\description{
Generates a file name in a specific format, which doesn't conflict with any files already in the folder.
}
