% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E004_nicePlot.R
\name{nicePlot}
\alias{nicePlot}
\title{nicePlot}
\usage{
nicePlot(
  art,
  ...,
  finalWidth = 3.6,
  finalHeight = 3.6,
  name = "artwork",
  folder = getwd(),
  bg = "WhiteSmoke",
  format = "png",
  auto = FALSE,
  bgCode = NULL,
  border = c(0.07, 0.03, 0),
  useGrid = FALSE,
  res = 300,
  innerMargin = 0.04,
  borderColours = c("white", "black"),
  borderSize = 60
)
}
\arguments{
\item{art}{a component, list of components....}

\item{finalWidth, finalHeight}{sizes given in inches.}

\item{name, folder, bg, format, border, useGrid, res, innerMargin, borderColours, borderSize, ...}{all passed to easyPlot}
}
\value{
No return value, easy plot is called.
}
\description{
A wrapper for easyPlot with defaults set up to make png files sized for instagram
}
