% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D009_p2pcf.R
\docType{class}
\name{p2pcf}
\alias{p2pcf}
\alias{initialize,p2pcf-method}
\alias{p2p_cf}
\alias{+,component,p2pcf-method}
\alias{-,component,p2pcf-method}
\title{p2pcf}
\usage{
\S4method{initialize}{p2pcf}(.Object, ..., functionList = as.character())

p2p_cf(functions = c(), ...)

\S4method{+}{component,p2pcf}(e1, e2)

\S4method{-}{component,p2pcf}(e1, e2)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing component functions that accept parameters.
}
\details{
This class is a convenient way of adding functions to component objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
