% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D011_p2plf.R
\docType{class}
\name{p2plf}
\alias{p2plf}
\alias{initialize,p2plf-method}
\alias{p2p_lf}
\alias{+,layer,p2plf-method}
\alias{-,layer,p2plf-method}
\title{p2plf}
\usage{
\S4method{initialize}{p2plf}(.Object, ..., functionList = as.character())

p2p_lf(functions = c(), ...)

\S4method{+}{layer,p2plf}(e1, e2)

\S4method{-}{layer,p2plf}(e1, e2)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing layer functions that accept parameters.
}
\details{
This class is a convenient way of adding functions to layer objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
