% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C008_pageFunctions.R
\name{pagefunctions}
\alias{pagefunctions}
\alias{runPageFunctions}
\title{pagefunctionsfunctions}
\usage{
runPageFunctions(
  input,
  classes = list(),
  getData = FALSE,
  runTimes = 0,
  frameNumber,
  plotting = FALSE
)
}
\description{
An function for running all the functions in a page
}
\examples{
print(1+1)
}
