% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C007_pages.R
\docType{class}
\name{pages}
\alias{pages}
\alias{scene}
\title{pages}
\usage{
scene(...)
}
\arguments{
\item{sections, ...}{These parameters are merged into the sections slot.}

\item{deviceSettings, }{name, style, p, units, frames, framesToPlot parameters for the constructor new("page") go directly into the relevant slots}
}
\description{
An S4 class to collect sections. It can represent a page in a pdf.;
}
\details{
The default page...
}
\section{Slots}{

\describe{
\item{\code{sections}}{A list of sections contained in the page}

\item{\code{name}}{A name for the file if the artwork is plotted in mode 2}

\item{\code{deviceSettings}}{A list of other device settings not covered by existing slots.}

\item{\code{style}}{A list of parameters inherited by all sections, layers and components in the page}

\item{\code{p}}{A list of parameters inherited by all sections, layers and components in the page}

\item{\code{units}}{A list of units inherited by all sections, layers and components in the page}

\item{\code{frames}}{numeric vector. The set of frames}

\item{\code{framesToPlot}}{numeric vector. The set of frames plotting functions can be run.}
}}

\examples{
print(1+1)
}
