% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J001_patternBasics.R
\name{patternBasics}
\alias{patternBasics}
\alias{blankPattern}
\alias{addLineToPattern}
\alias{addPolygonToPattern}
\title{Pattern Building Blocks}
\usage{
blankPattern(x0, x1, y0, y1, border = "black", background = "white", ...)

addLineToPattern(pattern, x0, x1, y0, y1, colour)

addPolygonToPattern(pattern, x, y, border, colour)
}
\arguments{
\item{x0, x1, y0, y1}{co-ordinates for the top, bottom, left and right of the square.}

\item{border}{colour for border lines. see details.}

\item{background}{the background colour for the square.}

\item{...}{parameters passed to component().}

\item{pattern}{a pattern to add a line or polygon to.}

\item{colour}{the colour of a line added to the pattern.}

\item{x, y}{co-ordinates for a line added to the pattern.}
}
\value{

}
\description{
Functions to draw a small square in a pattern step by step.
}
\details{
blankPattern creates a component with a single square, and the other functions add details.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{checkPatterns}},
\code{\link{makePatternBlock}()},
\code{\link{patternNames}()},
\code{\link{squarePatterns}},
\code{\link{stripePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
