% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J003_patternNames.R
\name{patternNames}
\alias{patternNames}
\title{patternNames}
\usage{
patternNames()
}
\value{
Returns a vector of names.
}
\description{
Contains all valid patternNames for makePatternBlock.
}
\examples{
patternNames()
}
\seealso{
Other patterns: 
\code{\link{checkPatterns}},
\code{\link{makePatternBlock}()},
\code{\link{patternBasics}},
\code{\link{squarePatterns}},
\code{\link{stripePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
