% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C011_plotting.R
\name{plottingArtObjects}
\alias{plottingArtObjects}
\alias{plotArtwork}
\alias{plotPageSingleFrame}
\alias{plotSection}
\alias{plotLayer}
\alias{plotComponent}
\title{plottingArtObjects}
\usage{
plotArtwork(art)

plotPageSingleFrame(input, data, units, classes, vp, useGrid, frameNumber)

plotSection(
  input,
  data = list(),
  units = list(),
  classes = as.character(),
  vp = list(x = 0, y = 0),
  useGrid = TRUE,
  frameNumber = 1
)

plotLayer(
  layer,
  useGrid,
  data = list(),
  units = list(),
  classes = as.character(),
  vp = list(x = 0, y = 0),
  xscale = 0:1,
  yscale = 0:1,
  frameNumber = 1
)

plotComponent(
  component,
  useGrid,
  data = list(),
  units = list(),
  classes = as.character(),
  vp = list(x = 0, y = 0),
  xscale = 0:1,
  yscale = 0:1
)
}
\arguments{
\item{art}{accepts only a single art object}
}
\value{

}
\description{
Functions to plot objects from this package.
}
\details{
plotArtwork, plotPage, plotSection, plotLayer and plotComponent are used for each art object.
Only plotArtwork is intended for end users.
plotArtwork's only input is an artwork object.
The others accept more parameters that come from the parent object.
}
\examples{
print(1+1)
}
