% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A006_polarConversions.R
\name{polar2cart}
\alias{polar2cart}
\title{polar2cart}
\usage{
polar2cart(x = 0, y = 0, distance, bearing, as.deg = FALSE)
}
\arguments{
\item{x, y}{the cartesian co-ordinates where angles are drawn from. Can be a vector and defaults to 0.}

\item{distance}{polar co-ordinate distance(s) from the center. Can be a vector}

\item{bearing}{polar co-ordinate angle(s). Can be a vector. By default this is in radians. See as.deg}

\item{as.deg}{defaults to false. Set to TRUE if you give bearings in degrees instead of radians.}
}
\value{
Returns a list with x and y co-ordinates.
}
\description{
Converts polar co-ordinates into cartesian co-ordinates.
}
\examples{
polar2cart(0,0,1,c(0,pi/2,pi,3*pi/2))
}
\seealso{
Other polarConversion: 
\code{\link{cart2polar}()},
\code{\link{makePolar}()}
}
\concept{polarConversion}
