% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A015_project.R
\name{project}
\alias{project}
\title{project}
\usage{
project(
  x,
  x0,
  x1,
  y,
  y0,
  y1,
  z,
  z0,
  z1,
  cameraXYZ = c(0, 0, 0),
  thetaXYZ = c(0, 0, 0),
  surfaceXYZ = c(0, 0, 0)
)
}
\arguments{
\item{x, y, z}{co-ordinates for points, polygons, etc}

\item{x0, x1, y0, y1, z0, z2}{co-ordiantes for lines}

\item{cameraXYZ}{co-ordinates for the camera}

\item{thetaXYZ}{direction of the camera}

\item{surfaceXYZ}{location of the surface}
}
\value{
Returns a list with x and y co-ordinates.
}
\description{
Converts 3d co-ordinates into 2d co-ordinates.
}
\details{
Uses conventions fromm physics. To Do: Add mathematics convention.
}
\examples{
print(1+1)
}
\concept{3Dgraphics}
