% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorManipulation.R
\name{pruneStart}
\alias{pruneStart}
\title{pruneStart}
\usage{
pruneStart(x, lengthToKeep)
}
\arguments{
\item{x}{an input vector}

\item{lengthToKeep}{determines the length of the output}
}
\value{
Returns a vector of a given length from the end of the input vector.
}
\description{
Takes a vector and returns the end of that vector.
}
\examples{
pruneStart(1:10,4)
}
