% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D009_quickFunctions.R
\name{quickfunctions}
\alias{quickfunctions}
\alias{action.data}
\alias{build.data}
\alias{camera.data}
\alias{display.data}
\alias{action.object}
\alias{build.object}
\alias{camera.object}
\alias{display.object}
\title{quickfunctions}
\usage{
action.data(fun = c(), .abcd = "a", .useData = TRUE, ...)

build.data(fun = c(), .abcd = "b", .useData = TRUE, ...)

camera.data(fun = c(), .abcd = "c", .useData = TRUE, ...)

display.data(fun = c(), .abcd = "d", .useData = TRUE, ...)

action.object(fun = c(), .abcd = "a", .useData = FALSE, ...)

build.object(fun = c(), .abcd = "b", .useData = FALSE, ...)

camera.object(fun = c(), .abcd = "c", .useData = FALSE, ...)

display.object(fun = c(), .abcd = "d", .useData = FALSE, ...)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing layer functions that accept layers.
}
\details{
This class is a convenient way of adding functions to layer objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
