% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A005_randomNumbers.R
\name{randomNumbers}
\alias{randomNumbers}
\alias{random}
\alias{randomMatrix}
\title{random number functions}
\usage{
random(lengthOut, maximum)

randomMatrix(numCol = 1, numRow = 1, maximum)
}
\arguments{
\item{lengthOut}{the length of the vector required.}

\item{maximum}{the largest possible random integer}

\item{numCol, }{numRow the dimensions of the matrix required. Both default to 1.}
}
\value{
random returns a vector of random numbers, whilst randomMatrix returns a matrix of random numbers.
}
\description{
Functions to return a either a vector or a matrix filled with random integers from 1 to maximum.
}
\examples{
random(10,5)
randomMatrix(3,6,10)
}
