% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A001_colourScheme.R
\name{RColorBrewerMaxColours}
\alias{RColorBrewerMaxColours}
\title{RColorBrewerMaxColours}
\usage{
RColorBrewerMaxColours(scheme = "RdPu")
}
\arguments{
\item{scheme}{The scheme in question}
}
\value{
The maximum number of colours for a given scheme in RColorBrewer
}
\description{
Returns the maximum number of colours for a given scheme in RColorBrewer
}
\examples{
RColorBrewerMaxColours=function(scheme="RdPu")
}
