\name{RGD-package}
\alias{RGD-package}
\alias{RGD}
\docType{package}
\title{
\packageTitle{RGD}
}
\description{
\packageDescription{RGD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RGD}
\packageIndices{RGD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RGD}

Maintainer: \packageMaintainer{RGD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
