% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C010_runningFunctions.R
\name{runningFunctions}
\alias{runningFunctions}
\alias{runAllFunctions}
\alias{cascadeFunction}
\title{runningFunctions}
\usage{
runAllFunctions(
  input,
  classes = list(),
  getData = FALSE,
  frameNumber,
  abcd = "a"
)

cascadeFunction(input, inputFunction, classes = list(), frameNumber)
}
\arguments{
\item{input}{a component, layer, section or page}

\item{classes}{list}

\item{getData}{logical}

\item{frameNumber}{numerical}

\item{abcd}{"a","b","c" or "d". Short for action, build, camera and display respectively.}
}
\value{

}
\description{
Functions to run all functions for art objects. This includes cascading a function down to objects contained in other objects.
}
\examples{
print(1+1)
}
