% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C006_sectionFunctions.R
\docType{class}
\name{sectionfunctions}
\alias{sectionfunctions}
\alias{initialize,sectionfunction-method}
\alias{runSectionFunction}
\alias{runSectionFunctions}
\title{sectionfunctions}
\usage{
\S4method{initialize}{sectionfunction}(
  .Object,
  fun = function(x) {     x },
  returnSection = TRUE,
  frames = integerSet()
)

runSectionFunction(input, inputFun, frameNumber)

runSectionFunctions(
  input,
  classes = list(),
  getData = FALSE,
  frameNumber = 1,
  plotting = FALSE
)
}
\description{
An S4 class for functions that operate on sections or their parameters
}
\details{
no convenient ways yet to add sectionfunctions to sections except manually.
}
\section{Slots}{

\describe{
\item{\code{fun}}{the name of a function or an function.}

\item{\code{returnSection}}{logical. If TRUE the function in fun accepts and returns sections. If FALSE it accepts parameters like a normal R function and returns a named list of parameters.}

\item{\code{frames}}{an integerSet object. Controls on which frames the function runs. The default behaviour is that it runs on all frames.}
}}

\examples{
print(1+1)
}
