% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C005_section.R
\docType{class}
\name{sections}
\alias{sections}
\alias{initialize,section-method}
\alias{section}
\title{sections}
\usage{
\S4method{initialize}{section}(
  .Object,
  ...,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  width = unit(1, "npc"),
  height = unit(1, "npc"),
  just = "centre",
  clip = "inherit",
  xscale = 0:1,
  yscale = 0:1,
  angle = 1,
  layers = list(),
  sections = list(),
  data = list(),
  style = list(),
  units = list(),
  borderCol = NULL,
  fillCol = NULL,
  borderLWD = 0.05,
  tableBox = NULL,
  action = list(),
  build = list(),
  camera = list(),
  display = list(),
  frames = integerSet(),
  tags = as.character()
)

section(...)
}
\arguments{
\item{x, y, width, height, clip, angle, xscale, yscale, sections, p, style, units, functions, tags}{parameters for the constructor new("section") go directly into the relevant slots}

\item{just}{can be 1 or 2 numbers, or a character vector including "left", "right", "top" or "bottom".}

\item{layers, ...}{These parameters are merged into the layers slot.}

\item{borderCol, fillCol, borderLWD}{REALLY NOT USED??? REMOVE}

\item{tableBox}{USED??????? REMOVE????????}

\item{frames}{If this is an integerSet, it goes directly into the relevant slot. If it is a numeric vector it is converted into an integerSet for you.}
}
\description{
An S4 class to represent a piece of a graphic. It can include a co-ordinate system.
}
\details{
The default layer made by the constructor completely empty, with no parameters, no components, no sublayers, no functions, no units. It is active and plotted on every frame.
}
\section{Slots}{

\describe{
\item{\code{x,y,just}}{used to position the section}

\item{\code{width,height}}{used to give the size of the section}

\item{\code{clip}}{if grid is used, this controls clipping.}

\item{\code{data}}{a list of parameters inherited by components, layers and sections in this section}

\item{\code{style}}{a list of parameters inherited by components, layers and sections in this section}

\item{\code{angle}}{used for rotation of viewports in grid}

\item{\code{xscale,yscale}}{used for co-ordinates that use "native" co-ordinate units}

\item{\code{layers}}{a list of layers}

\item{\code{sections}}{a list of sections}

\item{\code{units}}{a list of units inherited by components, layers and sections in this section}

\item{\code{action,build,display}}{a list of section functions that can apply to the section, the list of parameters in data or layers, components and sections contained in the section.}

\item{\code{camera}}{a list of functions that apply to layers and components in the section.}

\item{\code{tags}}{USED??? REMOVE??}

\item{\code{frames}}{controls which frames the section is plotted.}
}}

\examples{
print(1+1)
}
