% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H002_Sierpinski.R
\name{sierpinski}
\alias{sierpinski}
\alias{runSierpinski}
\alias{makeSierpinski}
\title{Sierpinski Triangle}
\usage{
runSierpinski(component)

makeSierpinski(type = "polygon", ...)
}
\arguments{
\item{component}{a triangle to be split up.}

\item{type}{the type of component}

\item{...}{parameters passed to the component and the component function. See "details" for more information.}
}
\value{
runSierpinski returns a list of components. makeSierpinski returns a component.
}
\description{
Functions to make sierpinski triangles.
}
\examples{
print(1+1)
}
\seealso{


Other fractals: 
\code{\link{spaceFillingCurve}}
}
\concept{fractals}
