% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E007_simpleAnimation.R
\name{simpleAnimation}
\alias{simpleAnimation}
\alias{parameterAdjust}
\alias{addSimpleAnimation}
\title{simpleAnimation}
\usage{
parameterAdjust(input)

addSimpleAnimation(input, parametersAdd, parametersMultiply, ...)
}
\arguments{
\item{input}{a component}

\item{parametersAdd}{a named list of parameters and changes. e.g. list(b=2,d=2) would add 2 to the value of b and d in input@data.}

\item{parametersMultiply}{a named list of parameters and scale factors. e.g. list(b=2,d=2) would double the value of b and d in input@data.}

\item{...}{function parameters for parameterAdjust from addSimpleAnimation.}
}
\value{
parameterAdjust and addSimpleAnimation both return components
}
\description{
Functions to create simple animations by varying parameters
}
\examples{
art=new("component",x=0,y=0)
art=addSimpleAnimation(art,list(x=0.01))
easyPlot(art,frames=1:100,framesToPlot=1:100,mode=0)
}
\seealso{

}
\concept{animation}
