% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H003_spaceFillingCurves.R
\name{spaceFillingCurve}
\alias{spaceFillingCurve}
\alias{makeSpaceFillingCurve}
\alias{runSpaceFillingCurve}
\title{Space Filling Curves}
\usage{
makeSpaceFillingCurve(type = "segments", ...)

runSpaceFillingCurve(
  x0,
  y0,
  x1,
  y1,
  percentDistances,
  bearing,
  addFinalSegment = TRUE,
  lineFlip = lineFlip,
  segmentsToFlip = c(),
  segmentOrder = c()
)
}
\arguments{
\item{...}{parameters passed to the component and the component function runSpaceFillingCurve. See "details" for more information.}
}
\value{
make returns a component, run returns a list.
}
\description{
Functions to calculate the sequence of points given by the following dynamical system in order to draw the attractor of the system.
}
\examples{
print(1+1)
}
\seealso{


Other fractals: 
\code{\link{sierpinski}}
}
\concept{fractals}
