% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A014_sphericalConversions.R
\name{spherical2cart}
\alias{spherical2cart}
\title{spherical2cart}
\usage{
spherical2cart(
  x = 0,
  y = 0,
  z = 0,
  radius,
  inclination,
  azimuth,
  as.deg = FALSE
)
}
\arguments{
\item{x, y}{the cartesian co-ordinates where angles are drawn from. Can be a vector and defaults to 0.}

\item{as.deg}{defaults to false. Set to TRUE if you give bearings in degrees instead of radians.}

\item{distance}{polar co-ordinate distance(s) from the center. Can be a vector}

\item{bearing}{polar co-ordinate angle(s). Can be a vector. By default this is in radians. See as.deg}
}
\value{
Returns a list with x, y and z co-ordinates.
}
\description{
Converts spherical co-ordinates into cartesian co-ordinates.
}
\details{
Uses conventions fromm physics. To Do: Add mathematics convention.
}
\examples{
print(1+1)
}
\concept{sphericalConversion}
