% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J006_squarePatterns.R
\name{squarePatterns}
\alias{squarePatterns}
\alias{squarePattern1}
\alias{squarePattern2}
\title{Square Patterns}
\usage{
squarePattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

squarePattern2(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)
}
\arguments{
\item{x0, x1, y0, y1}{co-ordinates for the top, bottom, left and right of the square.}

\item{backgroundBorder, patternBorder}{colour for the lines in the background and foreground.}

\item{foreground, background}{colours for the foreground and background.}

\item{...}{parameters passed to the component.}
}
\value{
return a single component
}
\description{
Functions to create square patterns
}
\details{
Two functions that make square patterns.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{checkPatterns}},
\code{\link{makePatternBlock}()},
\code{\link{patternBasics}},
\code{\link{patternNames}()},
\code{\link{stripePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
