% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K001_stringArtObject.R
\docType{class}
\name{stringArt}
\alias{stringArt}
\alias{initialize,stringArt-method}
\title{An S4 class to represent a piece of string art}
\usage{
\S4method{initialize}{stringArt}(.Object, x, y, names, groups, lineGroups, distances, loop)
}
\description{
An S4 class to describe a piece of string art.
}
\section{Slots}{

\describe{
\item{\code{x,y}}{point co-ordinates}

\item{\code{names}}{point names}

\item{\code{groups}}{a named list of name vectors}

\item{\code{distances}}{a list of adjacencies for each line group}

\item{\code{lineGroups}}{a list of point groups to join with lines}

\item{\code{loop}}{logical. If true the ends of a group are considered adjacent to the start of the group.}
}}

\examples{
print(1+1)
}
\seealso{
Other stringArt: 
\code{\link{makeMandalaStringArt}()},
\code{\link{makeShapeStringArt}()},
\code{\link{makeStringArt}()}
}
\concept{stringArt}
