% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J005_stripePatterns.R
\name{stripePatterns}
\alias{stripePatterns}
\alias{diagonalStripePattern1}
\alias{verticalStripePattern1}
\alias{horizontalStripePattern1}
\title{Stripe Patterns}
\usage{
diagonalStripePattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

verticalStripePattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

horizontalStripePattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)
}
\arguments{
\item{x0, x1, y0, y1}{co-ordinates for the top, bottom, left and right of the square.}

\item{backgroundBorder, patternBorder}{colour for the lines in the background and foreground.}

\item{foreground, background}{colours for the foreground and background.}

\item{...}{parameters passed to the component.}
}
\value{
return a single component
}
\description{
Functions to create striped patterns
}
\details{
Three functions that make striped patterns.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{checkPatterns}},
\code{\link{makePatternBlock}()},
\code{\link{patternBasics}},
\code{\link{patternNames}()},
\code{\link{squarePatterns}},
\code{\link{trianglePatterns}}
}
\concept{patterns}
