% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A004_subsetHelpers.R
\name{subsetHelpers}
\alias{subsetHelpers}
\alias{loopSubset}
\alias{overflowSubset}
\title{Subset helper functions}
\usage{
loopSubset(x, i)

overflowSubset(x, i)
}
\arguments{
\item{x}{a vector to be subsetted.}

\item{i}{a vector of indexes}
}
\value{
A vector containing elements of x, chosen by the index i.
}
\description{
Functions that return x\verb{[i}] but with alternate behaviour when i is greater than the length of x.
}
\details{
loopSubset will cycle back through the elements of x when i is too big. overflowSubset will simply give the last element of x instead.
}
\examples{
loopSubset(1:10,1:20)
overflowSubset(1:10,1:20)
}
