% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D003_translateObjects.R
\name{translateObjects}
\alias{translateObjects}
\alias{translateComponent}
\alias{translateLayer}
\title{translateObjects}
\usage{
translateComponent(
  input,
  translateX = 0,
  translateY = 0,
  xcoordinates = coordinateNamesX(),
  ycoordinates = coordinateNamesY()
)

translateLayer(
  input,
  translateX = 0,
  translateY = 0,
  xcoordinates = coordinateNamesX(),
  ycoordinates = coordinateNamesY()
)
}
\arguments{
\item{input}{a component or layer to translate}

\item{translateX, translateY}{numeric. An amount to translate}

\item{xcoordinates}{a list of names of parameters considered to be x co-ordinates}

\item{ycoordinates}{a list of names of parameters considered to be y co-ordinates}
}
\value{
returns a modified version of the input
}
\description{
Functions to manipulate co-ordinate parameters.
}
\examples{
print(1+1)
}
