% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J004_trianglePatterns.R
\name{trianglePatterns}
\alias{trianglePatterns}
\alias{trianglePattern1}
\alias{trianglePattern2}
\alias{trianglePattern3}
\title{Triangle Patterns}
\usage{
trianglePattern1(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

trianglePattern2(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)

trianglePattern3(
  x0,
  x1,
  y0,
  y1,
  backgroundBorder = "black",
  background = "black",
  patternBorder = "black",
  foreground = "white",
  ...
)
}
\arguments{
\item{x0, x1, y0, y1}{co-ordinates for the top, bottom, left and right of the square.}

\item{backgroundBorder, patternBorder}{colour for the lines in the background and foreground.}

\item{foreground, background}{colours for the foreground and background.}

\item{...}{parameters passed to the component.}
}
\value{
return a single component
}
\description{
Functions to create triangular patterns
}
\details{
Three functions that make triangular patterns.
}
\examples{
print(1+1)
}
\seealso{


Other patterns: 
\code{\link{checkPatterns}},
\code{\link{makePatternBlock}()},
\code{\link{patternBasics}},
\code{\link{patternNames}()},
\code{\link{squarePatterns}},
\code{\link{stripePatterns}}
}
\concept{patterns}
