% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B006_validParameters.R
\name{validPars}
\alias{validPars}
\alias{validParameters}
\alias{validAdjustParameters}
\alias{coordinateNamesX}
\alias{coordinateNamesY}
\title{Valid Parameters}
\usage{
validParameters(fun = "", simplify = TRUE, style = FALSE, useGrid = FALSE)

validAdjustParameters()

coordinateNamesX()

coordinateNamesY()
}
\arguments{
\item{fun}{a function}

\item{simplify}{logical}

\item{style}{logical}

\item{useGrid}{logical}
}
\value{
a vector of parameter names
}
\description{
\itemize{
\item validParameters calculates valid parameters for any function, or for par and gpar.
\item validAdjustParameters returns a list of parameters that can modify co-ordinate parameters
\item coordinateNamesX returns a list of parameters considered to be x co-ordinates
\item coordinateNamesY returns a list of parameters considered to be x co-ordinates
}
}
\examples{
print(1+1)
}
