% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A001_colourScheme.R
\name{wesandersonMaxColours}
\alias{wesandersonMaxColours}
\title{wesandersonMaxColours}
\usage{
wesandersonMaxColours(scheme = "GrandBudapest1")
}
\arguments{
\item{scheme}{The scheme in question}
}
\value{
The maximum number of colours for a given scheme in wes anderson
}
\description{
Returns the maximum number of colours for a given scheme in the wes anderson package
}
\examples{
wesandersonMaxColours=function(scheme="GrandBudapest1")
}
