% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A018_antiAlias.r
\name{antiAlias22}
\alias{antiAlias22}
\title{antiAlias}
\usage{
antiAlias22(image)
}
\arguments{
\item{image}{an array with three colour channels}
}
\value{
Returns an array that is half the size of the original
}
\description{
A simple anti alias function
}
\examples{
print(1+1)
}
