% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A001_colourScheme.R, R/colourScheme.R
\name{colourScheme}
\alias{colourScheme}
\title{colourScheme}
\usage{
colourScheme(
  numberOfColours = 10,
  style = "BW",
  colours = NULL,
  bias = 1,
  space = c("rgb", "Lab"),
  alpha = FALSE,
  interpolate = c("linear", "spline"),
  reverse = FALSE,
  discrete = TRUE
)

colourScheme(
  numberOfColours = 10,
  style = "BW",
  colours = NULL,
  bias = 1,
  space = c("rgb", "Lab"),
  alpha = FALSE,
  interpolate = c("linear", "spline"),
  reverse = FALSE,
  discrete = TRUE
)
}
\arguments{
\item{numberOfColours}{The number of colours returned. Defaults to 10.}

\item{style}{The palette or function to use. Used when colours is NULL. Defaults to "BW" for black and white.}

\item{colours}{A starting set of colours that will be interpolated.}

\item{bias}{used by colorRampPalette}

\item{space}{used by colorRampPalette}

\item{alpha}{used by colorRampPalette}

\item{interpolate}{used by colorRampPalette}

\item{reverse}{Reverses the final colour order. Defaults to FALSE}

\item{discrete}{Defaults to TRUE}
}
\value{
Returns a vector of colours of the given length.

Returns a vector of colours of the given length.
}
\description{
Creates a single function that can create colour schemes using many different packages with the option of interpolation.

Creates a single function that can create colour schemes using many different packages with the option of interpolation.
}
\details{
The possible values of style are;
\itemize{
\item "BW" and "BWrev" for greyscale
\item "rainbow" and "rainbowrev" which uses rainbow()
\item "heat" and "heatrev" which uses heat.colors()
\item "terrain" and "terrainrev" which uses terrain.colors()
\item "topo" and "toporev" which uses topo.colors()
\item "cm" and "cmrev" which uses cm.colors()
\item "BrBG","PiYG","PRGn","PuOr","RdBu","RdGy","RdYlBu",
"RdYlGn","Spectral","Accent","Dark2","Paired","Pastel1","Pastel2","Set1",
"Set2","Set3","Blues","BuGn","BuPu","GnBu","Greens","Greys","Oranges",
"OrRd","PuBu","PuBuGn","PuRd","Purples","RdPu","Reds","YlGn","YlGnBu",
"YlOrBr" and "YlOrRd" from the RColorBrewer package
\item "plasma", "plasmarev", "viridis", "viridisrev", "magma", "magmarev", "inferno" and "infernorev" from the viridis package.
}

*"BottleRocket1","BottleRocket2","Rushmore1","Rushmore",
"Royal1","Royal2","Zissou1","Darjeeling1",
"Darjeeling2","Chevalier1","FantasticFox1","Moonrise1",
"Moonrise2","Moonrise3","Cavalcanti1","GrandBudapest1",
"GrandBudapest2","IsleofDogs1" and "IsleofDogs2" from the wesanderson package.

The above terms that end in rev overrides the reverse parameter to force reversing the colour scheme.
}
\examples{
colourScheme()
colourScheme(25,"rainbow")
colourScheme()
colourScheme(25,"rainbow")
}
