% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C001_component.R
\name{components}
\alias{components}
\alias{component}
\alias{initialize,component-method}
\title{components}
\usage{
component(...)

\S4method{initialize}{component}(
  .Object,
  ...,
  type = "points",
  style = list(),
  units = list(),
  data = list(),
  action = list(),
  build = list(),
  display = list(),
  active = TRUE,
  visible = TRUE,
  ref = as.character(),
  frames = integerSet()
)

component(...)
}
\arguments{
\item{type, style, units, action, build, display, active, visible}{parameters for the constructor new("component") go directly into the relevant slots}

\item{data, ref, ...}{These parameters are merged into the data slot. Note any arguments named ".type",".frames", ".abcd", ".useData" and ".repeats" are ignored.}

\item{frames}{If this is an integerSet, it goes directly into the relevant slot. If it is a numeric vector it is converted into an integerSet for you.}
}
\description{
An S4 class to represent a single element in a graphic;
}
\details{
The default component made by the constructor is a point, with no parameters, no functions and no units. It is active and plotted on every frame.
}
\section{Slots}{

\describe{
\item{\code{type}}{The name(s) of the plotting function(s) used to plot the component. Defaults to "points". This can also include the names of classes linked to a plotting function.}

\item{\code{units}}{A list of units to use for parameters.}

\item{\code{data}}{a list of parameters.}

\item{\code{style}}{a list of parameters to pass to par or gpar that are set for this component only.}

\item{\code{active}}{logical. If FALSE the component functions aren't run}

\item{\code{visible}}{logical. If FALSE the component isn't plotted.}

\item{\code{action}}{list. Functions that act on the component or the elements of data. Changes made persist between frames. Actions run before build, camera, display and plot functions.}

\item{\code{build}}{list. Functions that act on the component or the elements of data. Changes made last until the frame is plotted. Build functions run before camera, display and plot functions.}

\item{\code{display}}{list. Functions that act on the component or the elements of data. Changes made last until the frame is plotted. Display functions run before plot functions.}

\item{\code{frames}}{an integerSet. The set of frames the component can be displayed in, (assuming visible is TRUE).}
}}

\examples{
print(1+1)
}
