% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A012_imageColourScheme.R
\name{imageColourScheme}
\alias{imageColourScheme}
\title{imageColourScheme}
\usage{
imageColourScheme(input)
}
\arguments{
\item{input}{a component}
}
\value{
Returns the input after setting input@p$matrixToColours_col
}
\description{
Takes a component and calculates a colour scheme for matrixToColours
}
\details{
imageColourScheme looks for arguments for colourScheme inside input@p
}
\examples{
imageColourScheme(component())@p
}
