% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M003_harmonograph.R
\name{makeHarmonograph}
\alias{makeHarmonograph}
\title{makeHarmonograph}
\usage{
makeHarmonograph(
  frequencies = c(3, 0, 4, 0),
  phases = c(pi/2, 0, 0, 0),
  amplitudes = c(1, 0, 1, 0),
  decay = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{frequencies}{pendulum paramater - numerical vector}

\item{phases}{pendulum paramater - numerical vector}

\item{amplitudes}{pendulum paramater - numerical vector}

\item{decay}{pendulum paramater - numerical vector}

\item{...}{parameters passed to makeParametric.}
}
\description{
Functions to calculate a harmonograph
runHarmonograph calculates the points. makeHarmonograph makes a component that calculates and plots the output.
}
\details{
The four inputs (frequencies, phases, amplitudes and decay) should all be vectors of the same length.
}
\examples{
easyPlot(makeHarmonograph())
}
\concept{makeParametric
easyPlot(makeHarmonograph())}
