% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A011_matrixToColours.R
\name{matrixToColours}
\alias{matrixToColours}
\title{matrixToColours}
\usage{
matrixToColours(
  image,
  col = list("white", "black", "red", "blue", "orange", "yellow", "green", "purple",
    "brown", "gray"),
  breaks = NULL
)
}
\arguments{
\item{image}{a matrix}

\item{col}{a list of colours}

\item{breaks}{optional. For ranges that correspond to colours}
}
\value{
Returns an RGB array
}
\description{
Takes a matrix converts into rgb colours. e.g. 1=red, 2=blue, etc or ranges for each colour
}
\examples{
matrixToColours(matrix(1:9,3))
}
