% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D009_quickFunctions.R
\name{quickfunctions}
\alias{quickfunctions}
\alias{action.data}
\title{quickfunctions}
\usage{
action.data(fun = c(), .abcd = "a", .useData = TRUE, ...)
}
\arguments{
\item{...}{extra items to be added to the functionList slot}

\item{functionList}{a list passed to the functionList slot}
}
\description{
An S4 class for storing layer functions that accept layers.
}
\details{
This class is a convenient way of adding functions to layer objects.
}
\section{Slots}{

\describe{
\item{\code{functionList}}{a list of functions. These can be names or actual functions.}
}}

\examples{
print(1+1)
}
