% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A001_colourScheme.R, R/colourScheme.R
\name{RColorBrewerMaxColours}
\alias{RColorBrewerMaxColours}
\title{RColorBrewerMaxColours}
\usage{
RColorBrewerMaxColours(style = "RdPu")

RColorBrewerMaxColours(style = "RdPu")
}
\arguments{
\item{style}{The style in question}
}
\value{
The maximum number of colours for a given style in RColorBrewer

The maximum number of colours for a given style in RColorBrewer
}
\description{
Returns the maximum number of colours for a given style in RColorBrewer

Returns the maximum number of colours for a given style in RColorBrewer
}
\examples{
RColorBrewerMaxColours=function(style="RdPu")
RColorBrewerMaxColours=function(style="RdPu")
}
