% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L002_svensson.R
\name{SvenssonAttractor}
\alias{SvenssonAttractor}
\alias{makeSvenssonAttractor}
\alias{runSvenssonAttractor}
\title{Johnny Svensson Attractors}
\usage{
makeSvenssonAttractor(type = "points", ...)

runSvenssonAttractor(
  a = 1.4,
  b = 1.56,
  c = 1.4,
  d = -6.56,
  inputx = 0,
  inputy = 0,
  x = NULL,
  y = NULL,
  runTime = 1e+05,
  sampleSize = NULL,
  reuseInput = FALSE
)
}
\arguments{
\item{type}{The name(s) of the plotting function(s) used to plot the component. Defaults to "points". This can also include the names of classes linked to a plotting function.}

\item{...}{parameters passed to the component and the component function runSvenssonAttractor. See "details" for more information.}

\item{a, b, c, d}{The four parameters of the attractor from the above equations.}
}
\description{
Functions to calculate the sequence of points given by the following dynamical system in order to draw the attractor of the system.
\itemize{
\item \ifelse{html}{\out{x[n+1]=d*sin(a*x[n])-sin(b*y[n])}}{\eqn{LATEX-LIKE CODE}{ASCII}}
\item \ifelse{html}{\out{y[n+1]=c*cos(a*x[n])+cos(b*y[n])}}{\eqn{LATEX-LIKE CODE}{ASCII}}
}

runSvenssonAttractor calculates the attractor. makeSvenssonAttractor makes a component that calculates and plots the Attractor.
}
\examples{
art=makeSvenssonAttractor(cex=0.1,col=rgb(0,0,0,0.1))
easyPlot(art,preserveAspectRatio=TRUE)
}
\seealso{
\url{http://paulbourke.net/fractals/peterdejong/}

Other 2dAttractors: 
\code{\link{CliffordAttractor}},
\code{\link{DeJongAttractor}}
}
\concept{2dAttractors}
