% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A001_colourScheme.R
\name{wesandersonMaxColours}
\alias{wesandersonMaxColours}
\title{wesandersonMaxColours}
\usage{
wesandersonMaxColours(style = "GrandBudapest1")
}
\arguments{
\item{style}{The style in question}
}
\value{
The maximum number of colours for a given style in wes anderson
}
\description{
Returns the maximum number of colours for a given style in the wes anderson package
}
\examples{
wesandersonMaxColours=function(style="GrandBudapest1")
}
